+++ crates/miden-lib/src/transaction/events.rs
@@ -54,7 +54,7 @@ const TX_SCRIPT_PROCESSING_START: u32 = 0x2_0016; // 131094
 const TX_SCRIPT_PROCESSING_END: u32 = 0x2_0017; // 131095
 
 const EPILOGUE_START: u32 = 0x2_0018; // 131096
-const EPILOGUE_TX_CYCLES_OBTAINED: u32 = 0x2_0019; // 131097
+const EPILOGUE_AFTER_TX_CYCLES_OBTAINED: u32 = 0x2_0019; // 131097
 const EPILOGUE_BEFORE_TX_FEE_REMOVED_FROM_ACCOUNT: u32 = 0x2_001a; // 131098
 const EPILOGUE_END: u32 = 0x2_001b; // 131099
 
@@ -62,6 +62,9 @@ const LINK_MAP_SET_EVENT: u32 = 0x2_001c; // 131100
 const LINK_MAP_GET_EVENT: u32 = 0x2_001d; // 131101
 
 const UNAUTHORIZED_EVENT: u32 = 0x2_001e; // 131102
+
+const EPILOGUE_AUTH_PROC_START: u32 = 0x2_0023; // 131107
+const EPILOGUE_AUTH_PROC_END: u32 = 0x2_0024; // 131108
 
 /// Events which may be emitted by a transaction kernel.
 ///
@@ -119,9 +122,13 @@ pub enum TransactionEvent {
     TxScriptProcessingEnd = TX_SCRIPT_PROCESSING_END,
 
     EpilogueStart = EPILOGUE_START,
-    EpilogueTxCyclesObtained = EPILOGUE_TX_CYCLES_OBTAINED,
-    EpilogueBeforeTxFeeRemovedFromAccount = EPILOGUE_BEFORE_TX_FEE_REMOVED_FROM_ACCOUNT,
     EpilogueEnd = EPILOGUE_END,
+
+    EpilogueAuthProcStart = EPILOGUE_AUTH_PROC_START,
+    EpilogueAuthProcEnd = EPILOGUE_AUTH_PROC_END,
+
+    EpilogueAfterTxCyclesObtained = EPILOGUE_AFTER_TX_CYCLES_OBTAINED,
+    EpilogueBeforeTxFeeRemovedFromAccount = EPILOGUE_BEFORE_TX_FEE_REMOVED_FROM_ACCOUNT,
 
     LinkMapSetEvent = LINK_MAP_SET_EVENT,
     LinkMapGetEvent = LINK_MAP_GET_EVENT,
@@ -214,7 +221,11 @@ impl TryFrom<u32> for TransactionEvent {
             TX_SCRIPT_PROCESSING_END => Ok(TransactionEvent::TxScriptProcessingEnd),
 
             EPILOGUE_START => Ok(TransactionEvent::EpilogueStart),
-            EPILOGUE_TX_CYCLES_OBTAINED => Ok(TransactionEvent::EpilogueTxCyclesObtained),
+            EPILOGUE_AUTH_PROC_START => Ok(TransactionEvent::EpilogueAuthProcStart),
+            EPILOGUE_AUTH_PROC_END => Ok(TransactionEvent::EpilogueAuthProcEnd),
+            EPILOGUE_AFTER_TX_CYCLES_OBTAINED => {
+                Ok(TransactionEvent::EpilogueAfterTxCyclesObtained)
+            },
             EPILOGUE_BEFORE_TX_FEE_REMOVED_FROM_ACCOUNT => {
                 Ok(TransactionEvent::EpilogueBeforeTxFeeRemovedFromAccount)
             },